1. Changes attributeEquals method to support order insensitive compare 2. Seperated labeled regex matcher test cases from regex matcher test cases and created new test class
