+++ textdb/textdb-exp/src/test/java/edu/uci/ics/textdb/exp/regexmatcher/RegexMatcherTest.java
@@ -388,126 +388,6 @@ public class RegexMatcherTest {
         Assert.assertEquals(exactResultsWithLimitOffset.size(), 2);
     }
 
-    @Test
-    public void testGetNextTupleLabeledRegex() throws Exception {
-        String query = "<name>";
-        String keywordQuery = "george lin lin";
-        List<Tuple> exactResults = RegexMatcherTestHelper.getQueryResults(
-                PEOPLE_TABLE, query, keywordQuery, Arrays.asList(TestConstants.FIRST_NAME), "name", false, Integer.MAX_VALUE, 0);
-
-        List<Tuple> expectedResults = new ArrayList<Tuple>();
-
-        // expected to match "george lin lin"
-        List<Tuple> data = TestConstants.getSamplePeopleTuples();
-        Schema spanSchema = Utils.addAttributeToSchema(TestConstants.SCHEMA_PEOPLE, new Attribute(RESULTS, AttributeType.LIST));
-        List<Span> spans = new ArrayList<Span>();
-        spans.add(new Span(TestConstants.FIRST_NAME, 0, 14, query, "george lin lin"));
-        IField spanField = new ListField<Span>(new ArrayList<Span>(spans));
-        List<IField> fields = new ArrayList<IField>(data.get(3).getFields());
-        fields.add(spanField);
-        expectedResults.add(new Tuple(spanSchema, fields.toArray(new IField[fields.size()])));
-
-        List<String> attributeNames = new ArrayList<String>();
-        attributeNames.add(RESULTS);
-        Assert.assertTrue(TestUtils.attributeEquals(expectedResults, exactResults, attributeNames));
-    }
-
-    @Test
-    public void testIgnoreCaseLabeledRegex() throws Exception {
-        String query = "<NAME>";
-        String keywordQuery = "george lin lin";
-        List<Tuple> exactResults = RegexMatcherTestHelper.getQueryResults(
-                PEOPLE_TABLE, query, keywordQuery, Arrays.asList(TestConstants.FIRST_NAME), "name", false, Integer.MAX_VALUE, 0);
-
-        List<Tuple> expectedResults = new ArrayList<Tuple>();
-
-        // expected to match "george lin lin"
-        List<Tuple> data = TestConstants.getSamplePeopleTuples();
-        Schema spanSchema = Utils.addAttributeToSchema(TestConstants.SCHEMA_PEOPLE, new Attribute(RESULTS, AttributeType.LIST));
-        List<Span> spans = new ArrayList<Span>();
-        spans.add(new Span(TestConstants.FIRST_NAME, 0, 14, query, "george lin lin"));
-        IField spanField = new ListField<Span>(new ArrayList<Span>(spans));
-        List<IField> fields = new ArrayList<IField>(data.get(3).getFields());
-        fields.add(spanField);
-        expectedResults.add(new Tuple(spanSchema, fields.toArray(new IField[fields.size()])));
-
-        List<String> attributeNames = new ArrayList<String>();
-        attributeNames.add(RESULTS);
-        Assert.assertTrue(TestUtils.attributeEquals(expectedResults, exactResults, attributeNames));
-    }
-
-    @Test
-    public void testEscapeCharacterLabeledRegex() throws Exception {
-        String query = "\\<name\\>";
-        String keywordQuery = "george lin lin";
-        List<Tuple> exactResults = RegexMatcherTestHelper.getQueryResults(
-                PEOPLE_TABLE, query, keywordQuery, Arrays.asList(TestConstants.FIRST_NAME), "name", false, Integer.MAX_VALUE, 0);
-        // exactResult should not contain any tuple
-        Assert.assertTrue(exactResults.size()==0);
-    }
-
-    @Test
-    public void testMultipleLabeledRegex() throws Exception {
-        String query = "<lab1>.*<lab2>";
-        List<Tuple> exactResults = RegexMatcherTestHelper.getQueryResults(
-                PEOPLE_TABLE, query, "short", Arrays.asList(TestConstants.DESCRIPTION), "lab1", false, Integer.MAX_VALUE, 0, "angry", "lab2");
-
-        List<Tuple> expectedResults = new ArrayList<Tuple>();
-
-        // expected to match "Short angry" and "Short and lin clooney is Angry"
-        List<Tuple> data = TestConstants.getSamplePeopleTuples();
-        Schema spanSchema = Utils.addAttributeToSchema(TestConstants.SCHEMA_PEOPLE, new Attribute(RESULTS, AttributeType.LIST));
-        List<Span> spans = new ArrayList<Span>();
-        spans.add(new Span(TestConstants.DESCRIPTION, 0, 11, query, "Short angry"));
-        IField spanField = new ListField<Span>(new ArrayList<Span>(spans));
-        List<IField> fields = new ArrayList<IField>(data.get(5).getFields());
-        fields.add(spanField);
-        expectedResults.add(new Tuple(spanSchema, fields.toArray(new IField[fields.size()])));
-
-        spans.clear();
-        spans.add(new Span(TestConstants.DESCRIPTION, 15, 45, query, "Short and lin clooney is Angry"));
-        spanField = new ListField<Span>(new ArrayList<Span>(spans));
-        fields = new ArrayList<IField>(data.get(3).getFields());
-        fields.add(spanField);
-        expectedResults.add(new Tuple(spanSchema, fields.toArray(new IField[fields.size()])));
-
-        List<String> attributeNames = new ArrayList<String>();
-        attributeNames.add(RESULTS);
-        Assert.assertTrue(TestUtils.attributeEquals(expectedResults, exactResults, attributeNames));
-    }
-
-    @Test
-    public void testDisjunctionLabeledRegex() throws Exception {
-        String query = "<lab1>|<lab2>";
-        List<Tuple> exactResults = RegexMatcherTestHelper.getQueryResults(
-                PEOPLE_TABLE, query, "short", Arrays.asList(TestConstants.DESCRIPTION), "lab1", false, Integer.MAX_VALUE, 0, "angry", "lab2");
-
-        List<Tuple> expectedResults = new ArrayList<Tuple>();
-
-        // expected to match "Short" and "angry"
-        List<Tuple> data = TestConstants.getSamplePeopleTuples();
-        Schema spanSchema = Utils.addAttributeToSchema(TestConstants.SCHEMA_PEOPLE, new Attribute(RESULTS, AttributeType.LIST));
-        List<Span> spans = new ArrayList<Span>();
-        spans.add(new Span(TestConstants.DESCRIPTION, 0, 5, query, "Short"));
-        spans.add(new Span(TestConstants.DESCRIPTION, 6, 11, query, "angry"));
-        IField spanField = new ListField<Span>(new ArrayList<Span>(spans));
-        List<IField> fields = new ArrayList<IField>(data.get(3).getFields());
-        fields.add(spanField);
-        expectedResults.add(new Tuple(spanSchema, fields.toArray(new IField[fields.size()])));
-        // expected to match "Short" and "Angry"
-        spans.clear();
-        spans.add(new Span(TestConstants.DESCRIPTION, 15, 20, query, "Short"));
-        spans.add(new Span(TestConstants.DESCRIPTION, 40, 45, query, "Angry"));
-        spanField = new ListField<Span>(new ArrayList<Span>(spans));
-        fields = new ArrayList<IField>(data.get(5).getFields());
-        fields.add(spanField);
-        expectedResults.add(new Tuple(spanSchema, fields.toArray(new IField[fields.size()])));
-
-        List<String> attributeNames = new ArrayList<String>();
-        attributeNames.add(RESULTS);
-        Assert.assertTrue(TestUtils.attributeEquals(expectedResults, exactResults, attributeNames));
-    }
-
     // @Test
     // public void testRegexWithLimitProblem() throws Exception {
     // List<ITuple> data = RegexTestConstantsText.getSampleTextTuples();
